move the status from lib to Harmony
